+++ gui/__init__.py
@@ -1,6 +1,6 @@
 """Tonverk Elmulti Converter GUI package."""
 
-__version__ = "1.1.3"
+__version__ = "1.1.4"
 
 from .app import ElmconvApp
 from .strings import Strings